+++ src/EditLevelLayer.cpp
@@ -12,10 +12,11 @@ class $modify(EditLevelLayer) {
 			return false;
 
 	time_t now = time(0);
-    tm* localTime = localtime(&now);
+	tm timeinfo;
+	errno_t err = localtime_s(&timeinfo, &now);
 
-	const int hr = localTime->tm_hour;
-    const int min = localTime->tm_min;
+    const int hr = timeinfo.tm_hour;
+    const int min = timeinfo.tm_min;
 
 	const int endhrbugfix = 24;
 	const int endminbugfix = 0;